ref(sdk-crashes): Move build detection configs
Move building crash detection configs to same file as the config.
