+++ src/sentry/utils/sdk_crashes/sdk_crash_detection_config.py
@@ -1,7 +1,10 @@
 from enum import Enum, unique
-from typing import Optional
+from typing import Optional, Sequence
 
+import sentry_sdk
 from typing_extensions import TypedDict
+
+from sentry import options
 
 
 @unique
@@ -21,3 +24,36 @@ class SDKCrashDetectionConfig(TypedDict):
     sample_rate: float
     """The organization allowlist to detect crashes for. If None, all organizations are allowed."""
     organization_allowlist: Optional[list[int]]
+
+
+def build_sdk_crash_detection_configs() -> Sequence[SDKCrashDetectionConfig]:
+    configs = [
+        _build_config(sdk_name=SdkName.Cocoa, has_allowlist=False),
+        _build_config(sdk_name=SdkName.ReactNative, has_allowlist=True),
+    ]
+
+    return [config for config in configs if config is not None]
+
+
+def _build_config(sdk_name: SdkName, has_allowlist: bool) -> Optional[SDKCrashDetectionConfig]:
+    options_prefix = f"issues.sdk_crash_detection.{sdk_name.value}"
+
+    project_id = options.get(f"{options_prefix}.project_id")
+    if not project_id:
+        sentry_sdk.capture_message(f"{sdk_name.value} project_id is not set.")
+        return None
+
+    sample_rate = options.get(f"{options_prefix}.sample_rate")
+    if not sample_rate:
+        return None
+
+    organization_allowlist: Optional[list[int]] = None
+    if has_allowlist:
+        organization_allowlist = options.get(f"{options_prefix}.organization_allowlist")
+
+    return SDKCrashDetectionConfig(
+        sdk_name=sdk_name,
+        project_id=project_id,
+        sample_rate=sample_rate,
+        organization_allowlist=organization_allowlist,
+    )
